Moved internal implementation of providers in LoadBalancer class from Set to Map.
Done because get() method will simulate HTTP GET call (we will introduce delay in next update to simulate congestion for task 8). Due to that we cannot check the identifier of the provider using its get() method - that would cause huge delays in Main program. We are getting the names by using keys from the Map implementation).